fix: 좋아요 개수 0인 상태에서 좋아요 취소되는 오류 해결

+++ client/src/components/common/card/review_card/CardBack.jsx
@@ -52,6 +52,8 @@ export default function CardBack({
         await bookApi.postBookLikes(id);
         setBookLikeCount((prev) => prev + 1);
       } else {
+        if (likeCount === 0) return;
+
         await bookApi.deleteBookLikes(id);
         setBookLikeCount((prev) => prev - 1);
       }
