+++ client/src/components/common/card/review_card/CardFront.jsx
@@ -34,7 +34,7 @@ export default function CardFront({
   const userId = useSelector((state) => state.auth.memberId);
   const modalRef = useRef(null);
 
-  const navigate = useNavigate();
+  const navigate = useNavigate();  
 
   useEffect(() => {
     setIsLiked(like);
@@ -56,6 +56,8 @@ export default function CardFront({
         await reviewApi.postReviewLikes(id);
         setReviewLikeCount((prev) => prev + 1);
       } else {
+        if (likeCount === 0) return;
+
         await reviewApi.deleteReviewLikes(id);
         setReviewLikeCount((prev) => prev - 1);
       }
